Composable storage
